Cap nhat chuc nang lien quan den tang giam gia, so luong

+++ Service/IImportOrders_IngredientsService.cs
@@ -5,7 +5,7 @@ namespace CheeseBurger.Service
 {
     public interface IImportOrders_IngredientsService
     {
-		void CreateOrderDetail(int orderID, int ingrID, int qty);
+		void CreateOrderDetail(int orderID, int ingrID, float qty);
 		void DeleteOrderDetail(int orderID);
 		List<ImportLineDTO> GetAllLines(int orderID);
 	}
